pyproject.toml
[project]
name = "django-holiday-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "respx>=0.22.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]
